tests/test_cloudinary_service.py
# !!! У файлі cloudinary_service.py зміненено рядок 55 img.save(buffered)  # Видалено 'format='PNG'

import cloudinary.uploader
from pymasters.services.cloudinary_service import upload_photo_to_cloudinary, delete_photo_from_cloudinary, create_transformation_urls, generate_qr_code
import pytest
from unittest.mock import patch
from pymasters.services.cloudinary_service import transform_photo

@pytest.fixture(scope="module", autouse=True)
def mock_cloudinary():
    with patch('cloudinary.uploader.upload') as mock_upload, \
         patch('cloudinary.uploader.destroy') as mock_destroy:
        mock_upload.return_value = {"url": "http://res.cloudinary.com/demo/image/upload/sample.jpg"}
        mock_destroy.return_value = {}
        yield mock_upload, mock_destroy

def test_upload_photo_to_cloudinary(mock_cloudinary):
    file = "path/to/test/image.jpg"
    result = upload_photo_to_cloudinary(file)
    assert result == "http://res.cloudinary.com/demo/image/upload/sample.jpg"

def test_delete_photo_from_cloudinary(mock_cloudinary):
    photo_url = "http://res.cloudinary.com/demo/image/upload/sample.jpg"
    delete_photo_from_cloudinary(photo_url)
    mock_cloudinary[1].assert_called_with("sample")

def test_create_transformation_urls():
    photo_url = "http://res.cloudinary.com/demo/image/upload/sample.jpg"
    transformations = [
        {"width": 300, "height": 300, "crop": "fill"},
        {"width": 600, "height": 400, "crop": "fit"}
    ]

    with patch('qrcode.image.pil.PilImage.save', new_callable=lambda: lambda *args, **kwargs: None) as mock_save:
        results = create_transformation_urls(photo_url, transformations)

    assert len(results) == 2
    for result in results:
        assert "transformation_url" in result
        assert "qr_code_url" in result
        assert result["qr_code_url"].startswith("http://res.cloudinary.com/demo/image/upload/")

def test_generate_qr_code(mock_cloudinary):
    mock_cloudinary[0].return_value = {"url": "http://example.com/qr_code.png"}

    with patch('qrcode.image.pil.PilImage.save', new_callable=lambda: lambda *args, **kwargs: None):
        result = generate_qr_code("http://example.com")
        assert result == "http://example.com/qr_code.png"

def test_transform_photo():
    photo_url = "http://res.cloudinary.com/demo/image/upload/sample.jpg"
    transformation = "width_300,height_300,c_fill"

    with patch('pymasters.services.cloudinary_service.create_transformation_urls') as mock_create_transformation_urls:
        mock_create_transformation_urls.return_value = [
            {
                "transformation_url": f"{photo_url}_{transformation}",
                "qr_code_url": f"http://res.cloudinary.com/demo/image/upload/sample_qr_code.png"
            }
        ]
        result = transform_photo(photo_url, transformation)
        assert "transformation_url" in result
        assert "qr_code_url" in result
        assert result["transformation_url"] == f"{photo_url}_{transformation}"
        assert result["qr_code_url"] == "http://res.cloudinary.com/demo/image/upload/sample_qr_code.png"

    invalid_transformation = "invalid_transformation"
    with patch('pymasters.services.cloudinary_service.create_transformation_urls') as mock_create_transformation_urls:
        mock_create_transformation_urls.return_value = []
        result = transform_photo(photo_url, invalid_transformation)
        assert "error" in result
        assert result["error"] == "Transformation not found"


